+++ packages/components/modules/profiles/web/ProfileMembers/MembersList/styled.tsx
@@ -1,0 +1,9 @@
+import { Box, styled } from '@mui/material'
+
+export const EmailListItemContainer = styled(Box)(() => ({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+  cursor: 'pointer',
+  width: '100%',
+}))
