+++ packages/components/modules/profiles/web/ProfileMembers/MembersList/types.ts
@@ -2,8 +2,17 @@ import { FC } from 'react'
 
 import { LoadingStateProps } from '@baseapp-frontend/design-system/components/web/displays'
 
+import { AvatarProps } from '@mui/material'
+import { RefetchFnDynamic } from 'react-relay'
+
 import { UserMembersListFragment$key } from '../../../../../__generated__/UserMembersListFragment.graphql'
+import { UsersListFragment$data } from '../../../../../__generated__/UsersListFragment.graphql'
 import { MemberItemProps } from '../MemberItem/types'
+
+export type User = NonNullable<
+  NonNullable<NonNullable<UsersListFragment$data['users']>['edges'][number]>['node']
+>
+export type NewEmail = { email: string; isNewEmail: boolean }
 
 export interface MembersListProps {
   MemberItem?: FC<MemberItemProps>
@@ -12,4 +21,18 @@ export interface MembersListProps {
   LoadingState?: FC<LoadingStateProps>
   LoadingStateProps?: LoadingStateProps
   membersContainerHeight?: number
+}
+
+export interface AddMembersDialogProps {
+  isOpen: boolean
+  onClose: () => void
+  profileId?: string
+  refetchMembers: RefetchFnDynamic<any, UserMembersListFragment$key>
+  LoadingStateProps?: LoadingStateProps
+}
+
+export interface UserCardProps {
+  user: User | NewEmail
+  onRemove: () => void
+  avatarProps?: AvatarProps
 }
